app/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
numpy==1.26.4
joblib==1.4.0
python-dotenv==1.0.1
google-generativeai==0.7.2
Werkzeug==3.0.3
scikit-learn == 1.4.2